Refactor plural child component to be singular, and move map up a layer to parent. Still refactoring, nothing broken.
